requirements.txt
certifi==2020.6.20
chardet==3.0.4
cloudinary==1.21.1
dj-database-url==0.5.0
Django==2.2.13
django-appconf==1.0.4
django-cloudinary-storage==0.2.3
django-heroku==0.3.1
django-imagekit==4.0.2
django-stdimage==5.1.1
django-taggit==1.2.0
gunicorn==20.0.4
idna==2.9
pilkit==2.0
Pillow==7.1.2
progressbar2==3.51.3
psycopg2==2.7.6.1
python-utils==2.4.0
pytz==2020.1
requests==2.24.0
six==1.15.0
sqlparse==0.3.1
urllib3==1.25.9
whitenoise==5.1.0
